+++ docs/assets/precache-manifest.1d0e544ae2a5c4957e2176dbff0a8220.js
@@ -1,16 +1,16 @@
 self.__precacheManifest = [
   {
-    "url": "/assets/vendors.7b8cda372549c004a38a.js"
+    "url": "/assets/vendors.1a4ba291ad87b89d24d8.js"
   },
   {
     "revision": "5ea9e5ec691f9b5f3150ed63859abdbc",
     "url": "/assets/modernizr-bundle.js"
   },
   {
-    "url": "/assets/app.7b8cda372549c004a38a.js"
+    "url": "/assets/app.1a4ba291ad87b89d24d8.js"
   },
   {
-    "revision": "c8a32b31576069a69b818d6306cb3436",
+    "revision": "4bb42d7687cd7b3e72d98bad9f0b8d29",
     "url": "/assets/../index.html"
   }
 ];